ethereum-listener - refactored and modularized sync logic
